Refactoring: Sharing route responsibilities with controllers
